dom/src/threads/test/test_scopeOnerror.html

<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads
-->
<head>
  <title>Test for DOM Worker Threads</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

  var errors = [
    "This exception should make it to onerror",
    "This error should not make it back out"
  ];

  var worker = new Worker("scopeOnerror_worker.js");

  worker.onmessage = function(event) {
    ok(event.data == "Done");
    SimpleTest.finish();
  }

  worker.onerror = function(event) {
    is(event.target, worker);
    is(event.message, "uncaught exception: " + errors[0]);
    event.preventDefault();
  };

  for (var index in errors) {
    worker.postMessage(errors[index]);
  }

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
